requirements.txt
aiosqlite==0.21.0
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.11.0
anywidget==0.9.18
appdirs==1.4.4
arrow==1.4.0
asn1crypto==1.5.1
astor==0.8.1
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.69
botocore==1.40.69
brotli==1.2.0
build==1.3.0
CacheControl==0.14.3
cachetools==6.2.1
captcha==0.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
dateutils==0.6.12
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docutils==0.22.3
duckdb==1.4.1
email-validator==2.3.0
entrypoints==0.4
executing==2.2.1
extra-streamlit-components==0.1.81
Faker==37.12.0
fal_client==0.9.1
favicon==0.7.0
fb_streamlit_auth==0.0.2
filelock==3.20.0
firebase_admin==7.1.0
fonttools==4.60.1
fpdf2==2.8.5
fsspec==2025.10.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud==0.34.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.5.0
google-cloud-vision==3.11.0
google-crc32c==1.7.1
google-genai==1.52.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gw_dsl_parser==0.1.49.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-oauth==0.15.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
id==1.5.0
idna==3.11
importlib_resources==6.5.2
ipylab==1.0.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.16
jwcrypto==1.5.6
kanaries_track==0.0.5
keyring==25.7.0
kiwisolver==1.4.9
langchain==1.0.3
langchain-core==1.0.3
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
monotonic==1.6
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.10.2
networkx==3.5
nh3==0.3.2
numpy==2.3.4
oauth2client==4.1.3
oauthlib==3.3.1
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
psygnal==0.15.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.2
pygwalker==0.4.9.15
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyphen==0.17.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
quickjs==1.19.4
readme_renderer==44.0
referencing==0.37.0
regex==2025.11.3
replicate==1.0.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
segment-analytics-python==2.2.3
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.0.0
snowflake-snowpark-python==1.42.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlglot==27.29.0
sqlite-vec==0.1.6
st-annotated-text==4.0.2
st-modal==1.0.2
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.51.0
streamlit-antd-components==0.3.2
streamlit-audiorecorder==0.0.6
streamlit-authenticator==0.3.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-chatbox==1.1.13.post1
streamlit-cookies-manager==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-feedback==0.1.4
streamlit-firebase-auth==1.1.5
streamlit-google-auth==1.1.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-markdown==1.1.0
streamlit-notify==0.3.1
streamlit-oauth==0.1.14
streamlit-option-menu==0.4.0
streamlit-permalink==1.2.0
streamlit-shadcn-ui==0.1.19
streamlit-supabase-auth==1.0.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
twine==6.2.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
validators==0.35.0
wasmtime==38.0.0
watchdog==6.0.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.15
xxhash==3.6.0
zopfli==0.4.0
zstandard==0.25.0


